A Collection is really a PaginatedCollection
I was extremely confused by this today as adding new resources which are
NOT paginated - they all come back at once, in an Array. I was trying to
use `resource.collection` and they always had a `#count` of `0`. Upon
digging in I realized those API repsponses were NOT paginated, but the
`Collection` didn't make it clear that it was only for paginated
collections. So, I propose we rename it. :)